+++ app/webservice.py
@@ -66,11 +66,7 @@ async def asr(
         audio_file: UploadFile = File(...),
         encode: bool = Query(default=True, description="Encode audio first through ffmpeg"),
         output: Union[str, None] = Query(default="txt", enum=["txt", "vtt", "srt", "tsv", "json"]),
-        word_timestamps: bool = Query(
-            default=False,
-            description="World level timestamps",
-            include_in_schema=(True if ASR_ENGINE == "faster_whisper" else False)
-        )
+        word_timestamps: bool = Query(default=False, description="World level timestamps")
 ):
     result = transcribe(load_audio(audio_file.file, encode), task, language, initial_prompt, word_timestamps, output)
     return StreamingResponse(